Concept: Create Actions
The browser allows you to set up event listeners and respond to certain events. Redux does too -- but the events are called actions. These actions are payloads of information that you set up to describe any event in the application that should update that application’s state.

actionCreate.js shows how to create an action